cleaned up test file structure
